core/kernel spec recovery: trace_var, public_method, caller/rand/Rational/Float/Array/String edges (144F/85E → 97F/54E) (#1032)

+++ monoruby/src/globals/error.rs
@@ -1315,7 +1315,7 @@ pub enum MonorubyErrKind {
     Key(Option<(Value, Value)>),
     Fiber,
     StopIteration,
-    SystemExit(u8),
+    SystemExit(i64),
     Other(ClassId),
     MethodReturn(Value, Lfp),
     /// `break` escaping a block: carries the break value, the block's
